created helper.py and added some little fixes
