+++ api/schedule.py
@@ -3,9 +3,10 @@ contains all /schedule endpoints
 '''
 from flask import Blueprint, request, jsonify
 from c1algo1.scheduler import generate_schedule as c1alg1
-# from c1algo2 import forecast as c1alg2 << not working right now, algo2 needs to debug this
+#from c1algo2 import forecast as c1alg2 ## not working right now, algo2 needs to debug this
 from coursescheduler import generate_schedule as c2alg1
 from forecaster.forecaster import forecast as c2alg2
+from .helper import get_prof_array, get_empty_schedule
 from .dbconn import DB_CONN
 
 SCHEDULE_BP = Blueprint('schedule', __name__)
@@ -15,86 +16,31 @@ def hello():
     blah
     '''
     return 'Hello from Schedules'
-SCHEDULES = [
-    {
-        'uuid': '4e90ab30-c380-4034-acdb-238856a88df3',
-        'semester': 'Fall',
-        'year': '2022',
-        'schedule': [
-            {
-                'timeslot': 'MWF 10am-10:50am',
-                'course_code': 'CSC230',
-                'prof': 'Bill Bird',
-                'section': 'A1',
-                'capacity': 100,
-            },
-            {
-                'timeslot': 'MWF 10am-10:50am',
-                'course_code': 'CSC230',
-                'prof': 'Bill Bird',
-                'section': 'A2',
-                'capacity': 20,
-            },
-            {
-                'timeslot': 'MTh 10am-10:50am',
-                'course_code': 'CSC111',
-                'prof': 'Hausi Muller',
-                'section': 'A1',
-                'capacity': 100,
-            },
-            {
-                'timeslot': 'MTh 10am-10:50am',
-                'course_code': 'CSC111',
-                'prof': 'Hausi Muller',
-                'section': 'A2',
-                'capacity': 20,
-            },
-        ]
-    },
-    {
-        'uuid': '5e90ab30-c380-4034-acdb-238856a88df3',
-        'semester': 'Spring',
-        'year': '2022',
-        'schedule': [
-            {
-                'timeslot': 'MWF 11am-11:50am',
-                'course_code': 'CSC370',
-                'prof': 'Bill Bird',
-                'section': 'A1',
-                'capacity': 40,
-            },
-            {
-                'timeslot': 'MWF 11am-11:50am',
-                'course_code': 'CSC370',
-                'prof': 'Bill Bird',
-                'section': 'A2',
-                'capacity': 10,
-            },
-            {
-                'timeslot': 'MTh 2pm-2:50pm',
-                'course_code': 'SENG275',
-                'prof': 'Mike Zastre',
-                'section': 'A1',
-                'capacity': 100,
-            },
-            {
-                'timeslot': 'MTh 2pm-2:50pm',
-                'course_code': 'SENG275',
-                'prof': 'Mike Zastre',
-                'section': 'A2',
-                'capacity': 20,
-            },
-        ]
-    }
-]
-UUIDS = [schedule['uuid'] for schedule in SCHEDULES]
 
 @SCHEDULE_BP.route('/', methods=['GET'])
 def get_all_schedules():
     '''
     Return JSON object containing a list of schedules with their year, semester and id
     '''
-    return jsonify(SCHEDULES), 200
+    sql = f"""SELECT
+                *
+        FROM Schedule;"""
+    results = DB_CONN.select(sql)
+    return results, 200
+
+@SCHEDULE_BP.route('/<schedule_id>', methods=['GET'])
+def get_schedule(schedule_id):
+    '''
+    Return a schedule with a particular id
+    '''
+    sql = f"""SELECT
+                *
+        FROM Schedule
+        WHERE BIN_TO_UUID(id) = \'{schedule_id}\';"""
+    results = DB_CONN.select_one(sql)
+    if results == None:
+        return 'Schedule not found',404
+    return results, 200
 
 @SCHEDULE_BP.route('/company/<company_num>', methods=['GET'])
 def get_company_schedule(company_num):
@@ -104,9 +50,17 @@ def get_company_schedule(company_num):
     '''
     prof_array = get_prof_array()
     schedule = get_empty_schedule()
-    # print(prof_array)
     message = f'company {company_num} not recognized'
     status = 200
+    # input to Algo 2
+    # historicalData: HistoricalCourseOffering[]
+    # previousEnrolment: ProgramEnrolment
+    # schedule: Schedule
+    # algo2_schedule = c1alg2(historical_data, previous_enrolment, schedule)
+    # input schedule to algo 1
+    # final_schedule = c1alg1(prof_array, algo2_schedule)
+    # post schedule
+    # return schedule
     if company_num == '1':
         message = 'Algo 1: ' + c1alg1(None, None, None)
         # message += ' Algo 2: ' + c1alg2(None, None, None) << not working same as above
@@ -116,180 +70,13 @@ def get_company_schedule(company_num):
     else:
         status = 404
     return message, status
-def get_empty_schedule():
-    '''Creates and returns an empty schedule pre-populated with courses
-    '''
-    #TODO: Add year_req
-    #TODO: add non-eng classes and timeslots
-    schedule = {}
-    schedule['fall'] = get_course_offering('fall_req')
-    schedule['spring'] = get_course_offering('spring_req')
-    schedule['summer'] = get_course_offering('summer_req')
-    return schedule
-def get_course_offering(semester:str):
-    '''
-    get and return the list of courseOffering for a certain semester
-    PARAMETERS: 'spring_req' or 'summer_req' or 'fall_req'
-    '''
-    sql = f"""SELECT
-                    course_code,
-                    course_name,
-                    spring_req,
-                    summer_req,
-                    fall_req,
-                    spring_peng_req,
-                    summer_peng_req,
-                    fall_peng_req
-            FROM CourseOffering
-            WHERE {semester} = 1;"""
-    results = DB_CONN.select(sql, ['spring_peng_req','summer_peng_req', 'fall_peng_req'])
-    courses_json = results.get_json()
-    courses = []
-    for course in courses_json:
-        new_course = {}
-        new_course['code'] = course['course_code']
-        new_course['title'] = course['course_name']
-        peng_required = {}
-        peng_required['fall'] = course['fall_peng_req']
-        peng_required['spring'] = course['spring_peng_req']
-        peng_required['summer'] = course['summer_peng_req']
-        new_course['pengRequired'] = peng_required
-        course_section = {}
-        course_sections = [course_section]
-        course_offering = {}
-        course_offering['course'] = new_course
-        course_offering['sections'] = course_sections
-        courses.append(course_offering)
-    return courses
-def get_prof_array():
-    '''
-    Creates and returns an array of Professors
-    '''
-    # TODO: Add prof time preferences
-    sql = f"""SELECT
-                    BIN_TO_UUID(ProfessorAvailability.id) as id,
-                    BIN_TO_UUID(ProfessorAvailability.prof_id) as prof_id, 
-                    ProfessorAvailability.year, 
-                    ProfessorAvailability.num_relief,
-                    ProfessorAvailability.num_summer_courses, 
-                    ProfessorAvailability.num_fall_courses, 
-                    ProfessorAvailability.num_spring_courses,
-                    Professor.first_name,
-                    Professor.last_name,
-                    Professor.is_peng,
-                    Professor.is_teaching
-            FROM ProfessorAvailability
-            LEFT JOIN Professor
-            ON ProfessorAvailability.prof_id = Professor.id;"""
-    results = DB_CONN.select(sql, ['is_peng', 'is_teaching'])
-    my_json = results.get_json()
-    prof_array = []
-    
-    i = 0
-    for prof in my_json:
-        new_prof = {}
-        prof_avail_id = my_json[i]['id']
-        new_prof['id'] = my_json[i]['prof_id']
-        new_prof['name'] = f"{my_json[i]['first_name']} {my_json[i]['last_name']}"
-        new_prof['isPeng'] = my_json[i]['is_peng']
-        if(my_json[i]['is_teaching']):
-            new_prof['facultyType'] = 'TEACHING'
-            num_classes = 6
-        else:
-            new_prof['facultyType'] = 'RESEARCH'
-            num_classes = 3
-        num_classes -= my_json[i]['num_relief']
-        if(num_classes<0):
-            num_classes = 0
-        new_prof['teachingObligations'] = num_classes
-        preferred_courses_per_semester = {}
-        preferred_courses_per_semester['fall'] = my_json[i]['num_fall_courses']
-        preferred_courses_per_semester['spring'] = my_json[i]['num_spring_courses']
-        preferred_courses_per_semester['summer'] = my_json[i]['num_summer_courses']
-        new_prof['preferredCoursesPerSemester'] = preferred_courses_per_semester
-        if(preferred_courses_per_semester['fall'] == 0):
-            new_prof['preferredNonTeachingSemester'] = 'FALL'
-        elif(preferred_courses_per_semester['spring'] == 0):
-            new_prof['preferredNonTeachingSemester'] = 'SPRING'
-        elif(preferred_courses_per_semester['summer'] == 0):
-            new_prof['preferredNonTeachingSemester'] = 'SUMMER'
-        # now added course Preferences
-        sql = f"""SELECT
-                    ProfessorCoursePreference.will_to_teach,
-                    ProfessorCoursePreference.able_to_teach,
-                    CourseOffering.course_code
-            FROM ProfessorAvailability
-            LEFT JOIN ProfessorCoursePreference
-            ON ProfessorAvailability.id = ProfessorCoursePreference.prof_avail_id
-            LEFT JOIN CourseOffering
-            ON ProfessorCoursePreference.course_id = CourseOffering.id
-            WHERE ProfessorCoursePreference.prof_avail_id = UUID_TO_BIN(\"{prof_avail_id}\");"""
-        results = DB_CONN.select(sql)
-        course_preferences_json = results.get_json()
-        course_preferences = []
-        pref_index = 0
-        for pref in course_preferences_json:
-            new_course_pref = {}
-            new_course_pref['courseCode'] = pref['course_code']
-            new_course_pref['enthusiasmScore'] = get_score(pref['able_to_teach'], pref['will_to_teach'])
-            course_preferences.append(new_course_pref)
-            pref_index += 1
-        new_prof['coursePreferences'] = course_preferences
-        prof_array.append(new_prof)
-        i += 1
-    return prof_array
-def get_score(able_to_teach:str, will_to_teach:str):
-    '''
-    Given an able to teach enum and will to teach returns the numerical score 
-    '''
-    score = 0
-    if able_to_teach== 'WITH_EFFORT':
-        if will_to_teach == 'UNWILLING':
-            score = 20
-        elif will_to_teach == 'WILLING':
-            score = 40
-        elif will_to_teach == 'VERY_WILLING':
-            score = 100
-    elif able_to_teach == 'ABLE':
-        if will_to_teach == 'UNWILLING':
-            score = 39
-        elif will_to_teach == 'WILLING':
-            score = 78
-        elif will_to_teach == 'VERY_WILLING':
-            score = 195
-    return score
-@SCHEDULE_BP.route('<schedule_id>', methods=['GET'])
-def get_schedule(schedule_id):
-    '''
-    Return JSON object containing schedule
-    '''
-    if schedule_id not in UUIDS:
-        return 'couldn\'t find that schedule', 404
-    return jsonify(SCHEDULES[UUIDS.index(schedule_id)]), 200
 
-@SCHEDULE_BP.route('/', methods=['POST'])
-def add_schedule():
+@SCHEDULE_BP.route('/<schedule_id>', methods=['PUT'])
+def update_schedule(schedule_id):
     '''
-    Request contains JSON object containing a new schedule
-    Add a new schedule to the table of schedules
-    Returns new schedule’s id
+    Update the schedule 
     '''
-    return f'schedule id of new entry\n\nJSON object:\n{request.data}\n\n', 200
-
-@SCHEDULE_BP.route('/<schedule_id>/<course_id>', methods=['PUT'])
-def update_course_time(schedule_id, course_id):
-    '''
-    Update the time slot of a course in the schedule
-    '''
-    return f'update timeslot of course {course_id} in schedule {schedule_id}', 200
-
-@SCHEDULE_BP.route('/<schedule_id>/<course_id>/<prof_id>', methods=['PUT'])
-def update_course_prof(schedule_id, course_id, prof_id):
-    '''
-    Update the professor of a course in the existing schedule
-    '''
-    return f'update prof teaching course {course_id} to \
-        prof {prof_id} in schedule {schedule_id}', 200
+    return f'update schedule {schedule_id}', 200
 
 @SCHEDULE_BP.route('/<schedule_id>', methods=['DELETE'])
 def delete_schedule(schedule_id):
@@ -297,10 +84,3 @@ def delete_schedule(schedule_id):
     Deletes a schedule from the schedules table
     '''
     return f'deleted schedule {schedule_id}', 200
-
-@SCHEDULE_BP.route('/<schedule_id>/<course_id>', methods=['DELETE'])
-def delete_course_from_schedule(schedule_id, course_id):
-    '''
-    Deletes a course from the schedule
-    '''
-    return f'deleted course {course_id} from schedule {schedule_id}', 200
